+++ programs/drift/src/instructions/admin.rs
@@ -21,9 +21,9 @@ use crate::math::constants::{
     DEFAULT_LIQUIDATION_MARGIN_BUFFER_RATIO, FEE_POOL_TO_REVENUE_POOL_THRESHOLD,
     IF_FACTOR_PRECISION, INSURANCE_A_MAX, INSURANCE_B_MAX, INSURANCE_C_MAX,
     INSURANCE_SPECULATIVE_MAX, LIQUIDATION_FEE_PRECISION, MAX_CONCENTRATION_COEFFICIENT,
-    MAX_SQRT_K, MAX_UPDATE_K_PRICE_CHANGE, PERCENTAGE_PRECISION, QUOTE_SPOT_MARKET_INDEX,
-    SPOT_CUMULATIVE_INTEREST_PRECISION, SPOT_IMF_PRECISION, SPOT_WEIGHT_PRECISION, THIRTEEN_DAY,
-    TWENTY_FOUR_HOUR,
+    MAX_SQRT_K, MAX_UPDATE_K_PRICE_CHANGE, PERCENTAGE_PRECISION, PERCENTAGE_PRECISION_I64,
+    QUOTE_SPOT_MARKET_INDEX, SPOT_CUMULATIVE_INTEREST_PRECISION, SPOT_IMF_PRECISION,
+    SPOT_WEIGHT_PRECISION, THIRTEEN_DAY, TWENTY_FOUR_HOUR,
 };
 use crate::math::cp_curve::get_update_k_result;
 use crate::math::orders::is_multiple_of_step_size;
@@ -1212,6 +1212,7 @@ pub fn handle_update_spot_market_oracle(
     ctx: Context<AdminUpdateSpotMarketOracle>,
     oracle: Pubkey,
     oracle_source: OracleSource,
+    skip_invariant_check: bool,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
     msg!("updating spot market {} oracle", spot_market.market_index);
@@ -1227,10 +1228,17 @@ pub fn handle_update_spot_market_oracle(
         oracle
     )?;
 
+    validate!(
+        ctx.accounts.old_oracle.key == &spot_market.oracle,
+        ErrorCode::DefaultError,
+        "old oracle account info ({:?}) and spot market oracle ({:?}) must match",
+        ctx.accounts.old_oracle.key,
+        spot_market.oracle
+    )?;
+
     // Verify oracle is readable
     let OraclePriceData {
-        price: _oracle_price,
-        delay: _oracle_delay,
+        price: new_oracle_price,
         ..
     } = get_oracle_price(&oracle_source, &ctx.accounts.oracle, clock.slot)?;
 
@@ -1245,6 +1253,40 @@ pub fn handle_update_spot_market_oracle(
         spot_market.oracle_source,
         oracle_source
     );
+
+    let OraclePriceData {
+        price: old_oracle_price,
+        ..
+    } = get_oracle_price(
+        &spot_market.oracle_source,
+        &ctx.accounts.old_oracle,
+        clock.slot,
+    )?;
+
+    msg!(
+        "Oracle Price: {:?} -> {:?}",
+        old_oracle_price,
+        new_oracle_price
+    );
+
+    if !skip_invariant_check {
+        validate!(
+            new_oracle_price > 0,
+            ErrorCode::DefaultError,
+            "invalid oracle price, must be greater than 0"
+        )?;
+
+        let oracle_change_divergence = new_oracle_price
+            .safe_sub(old_oracle_price)?
+            .safe_mul(PERCENTAGE_PRECISION_I64)?
+            .safe_div(old_oracle_price)?;
+
+        validate!(
+            oracle_change_divergence.abs() < (PERCENTAGE_PRECISION_I64 / 10),
+            ErrorCode::DefaultError,
+            "invalid new oracle price, more than 10% divergence"
+        )?;
+    }
 
     spot_market.oracle = oracle;
     spot_market.oracle_source = oracle_source;
@@ -3431,9 +3473,10 @@ pub fn handle_update_state_max_initialize_user_fee(
     perp_market_valid(&ctx.accounts.perp_market)
 )]
 pub fn handle_update_perp_market_oracle(
-    ctx: Context<RepegCurve>,
+    ctx: Context<AdminUpdatePerpMarketOracle>,
     oracle: Pubkey,
     oracle_source: OracleSource,
+    skip_invariant_check: bool,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
     msg!("perp market {}", perp_market.market_index);
@@ -3450,9 +3493,17 @@ pub fn handle_update_perp_market_oracle(
         oracle
     )?;
 
-    // Verify oracle is readable
+    validate!(
+        ctx.accounts.old_oracle.key == &perp_market.amm.oracle,
+        ErrorCode::DefaultError,
+        "old oracle account info ({:?}) and perp market oracle ({:?}) must match",
+        ctx.accounts.old_oracle.key,
+        perp_market.amm.oracle
+    )?;
+
+    // Verify new oracle is readable
     let OraclePriceData {
-        price: _oracle_price,
+        price: new_oracle_price,
         delay: _oracle_delay,
         ..
     } = get_oracle_price(&oracle_source, &ctx.accounts.oracle, clock.slot)?;
@@ -3468,6 +3519,40 @@ pub fn handle_update_perp_market_oracle(
         perp_market.amm.oracle_source,
         oracle_source
     );
+
+    let OraclePriceData {
+        price: old_oracle_price,
+        ..
+    } = get_oracle_price(
+        &perp_market.amm.oracle_source,
+        &ctx.accounts.old_oracle,
+        clock.slot,
+    )?;
+
+    msg!(
+        "Oracle Price: {:?} -> {:?}",
+        old_oracle_price,
+        new_oracle_price
+    );
+
+    if !skip_invariant_check {
+        validate!(
+            new_oracle_price > 0,
+            ErrorCode::DefaultError,
+            "invalid oracle price, must be greater than 0"
+        )?;
+
+        let oracle_change_divergence = new_oracle_price
+            .safe_sub(old_oracle_price)?
+            .safe_mul(PERCENTAGE_PRECISION_I64)?
+            .safe_div(old_oracle_price)?;
+
+        validate!(
+            oracle_change_divergence.abs() < (PERCENTAGE_PRECISION_I64 / 10),
+            ErrorCode::DefaultError,
+            "invalid new oracle price, more than 10% divergence"
+        )?;
+    }
 
     perp_market.amm.oracle = oracle;
     perp_market.amm.oracle_source = oracle_source;
@@ -4731,6 +4816,23 @@ pub struct AdminUpdateSpotMarketOracle<'info> {
     pub spot_market: AccountLoader<'info, SpotMarket>,
     /// CHECK: checked in `initialize_spot_market`
     pub oracle: AccountInfo<'info>,
+    /// CHECK: checked in `admin_update_spot_market_oracle` ix constraint
+    pub old_oracle: AccountInfo<'info>,
+}
+
+#[derive(Accounts)]
+pub struct AdminUpdatePerpMarketOracle<'info> {
+    pub admin: Signer<'info>,
+    #[account(
+        has_one = admin
+    )]
+    pub state: Box<Account<'info, State>>,
+    #[account(mut)]
+    pub perp_market: AccountLoader<'info, PerpMarket>,
+    /// CHECK: checked in `admin_update_perp_market_oracle` ix constraint
+    pub oracle: AccountInfo<'info>,
+    /// CHECK: checked in `admin_update_perp_market_oracle` ix constraint
+    pub old_oracle: AccountInfo<'info>,
 }
 
 #[derive(Accounts)]